[State Sync] Move src/tests into tests directory to force integration test constraints.
Closes: #7290
