+++ state-synchronizer/src/lib.rs
@@ -7,19 +7,15 @@
 //! Used for node restarts, network partitions, full node syncs
 #![recursion_limit = "1024"]
 
-pub use self::state_synchronizer::{StateSyncBootstrapper, StateSyncClient};
-
 pub mod chunk_request;
 pub mod chunk_response;
 pub mod coordinator;
 mod counters;
-mod executor_proxy;
+pub mod executor_proxy;
 mod logging;
 pub mod network;
 mod request_manager;
-mod state_synchronizer;
+pub mod state_synchronizer;
 
 #[cfg(any(feature = "fuzzing", test))]
-mod tests;
-#[cfg(any(feature = "fuzzing", test))]
-pub use tests::fuzzing;
+pub use coordinator::fuzzing;
