--- state-synchronizer/src/tests/mod.rs
@@ -1,12 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-#[cfg(test)]
-mod integration_tests;
-#[cfg(test)]
-mod mock_executor_proxy;
-#[cfg(test)]
-mod mock_storage;
-
-#[cfg(any(feature = "fuzzing", test))]
-pub mod fuzzing;
